# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TNU Admission Chatbot API",
        "description": "API tài liệu hóa cho chatbot tư vấn tuyển sinh, dashboard admin và CMS dữ liệu tuyển sinh.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Laravel API"
        }
    ],
    "paths": {
...
        "/api/knowledge-bases": {
            "get": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Danh sach knowledge base",
                "operationId": "47e61c307a4996fd7ddeae803829a766",
                "responses": {
                    "200": {
                        "description": "Danh sach knowledge base"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Tao knowledge base",
                "operationId": "35d70b37c4c167e0c35fb92c96e2bbe8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "nganh_dao_tao"
                                    },
                                    "source_type": {
                                        "type": "string",
                                        "example": "nganh_dao_tao"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Cong nghe thong tin 2025"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Noi dung tuyen sinh..."
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tao thanh cong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/knowledge-bases/{knowledge_base}": {
            "get": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Chi tiet knowledge base",
                "operationId": "86c41cdc6511edddfa189e5233e0c630",
                "parameters": [
                    {
                        "name": "knowledge_base",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiet knowledge base"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Cap nhat knowledge base",
                "operationId": "a52783cb4915ae7b6d79e8134e4a70cc",
                "parameters": [
                    {
                        "name": "knowledge_base",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tieu de moi"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Noi dung moi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cap nhat thanh cong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Knowledge Base"
                ],
                "summary": "Xoa knowledge base",
                "operationId": "1bfa1050456b5228461ec43beb5292c3",
                "parameters": [
                    {
                        "name": "knowledge_base",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xoa thanh cong"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Nhập theo định dạng: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}